+++ src/js/util/lang.js
@@ -173,3 +173,9 @@ export function toMs(time) {
             ? parseFloat(time)
             : parseFloat(time) * 1000;
 }
+
+export function swap(value, a, b) {
+    return value.replace(new RegExp(`${a}|${b}`, 'mg'), function (match) {
+        return match === a ? b : a
+    });
+}