+++ webpack.config.js
@@ -24,9 +24,29 @@ module.exports = [
         },
         module: loaders,
         externals: {jquery: 'jQuery'},
+        resolve: {
+            alias: {
+                "components$": __dirname + "/dist/icons/components.json",
+            }
+        },
         plugins: [
             new BuildAll()
         ]
+    },
+
+    {
+        entry: './src/js/icons',
+        output: {
+            filename: 'dist/js/uikit-icons.js',
+            library: 'UIkitIcons',
+            libraryTarget: 'umd'
+        },
+        module: loaders,
+        resolve: {
+            alias: {
+                "icons$": __dirname + "/dist/icons/icons.json",
+            }
+        }
     },
 
     {